+++ clang-tools-extra/clang-tidy/hooks/HooksTidyModule.cpp
@@ -42,6 +42,7 @@
 #include "SlotHashBufLenCheck.h"
 #include "SlotKeyletBufLenCheck.h"
 #include "SlotLimitCheck.h"
+#include "SlotSubLimitCheck.h"
 #include "ValidateBufLenCheck.h"
 #include "VerifyBufLenCheck.h"
 
@@ -118,6 +119,8 @@ public:
         "hooks-slot-keylet-buf-len");
     CheckFactories.registerCheck<SlotLimitCheck>(
         "hooks-slot-limit");
+    CheckFactories.registerCheck<SlotSubLimitCheck>(
+        "hooks-slot-sub-limit");
     CheckFactories.registerCheck<ValidateBufLenCheck>(
         "hooks-validate-buf-len");
     CheckFactories.registerCheck<VerifyBufLenCheck>(